requirements.txt
annotated-types==0.7.0
anyio==4.12.1
asgiref==3.11.1
certifi==2026.2.25
colorama==0.4.6
distro==1.9.0
Django==6.0.3
django-admin-interface==0.32.0
django-ckeditor==6.7.3
django-colorfield==0.14.0
django-extensions==4.1
django-js-asset==3.1.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.13.0
mysqlclient==2.2.8
numpy==2.4.2
openai==2.28.0
opencv-python==4.13.0.92
pillow==12.1.1
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.2
python-slugify==8.0.4
sniffio==1.3.1
sqlparse==0.5.5
text-unidecode==1.3
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
